refactor(gui): preparation for node graph
